Extract BaseOcrActivity and re-implement VatInvoiceActivity by deriving it from BaseOcrActivity.
